Organizando arquivos de teste. Dividindo entre Paciente e Profissionais. [#175472150]
