+++ Saudox/tests/Feature/Profissional_AgendamentosTest.php
@@ -9,7 +9,7 @@ use App\Endereco;
 use Illuminate\Support\Facades\Auth;
 use Tests\TestCase;
 
-class AgendamentosTest extends TestCase {
+class Profissional_AgendamentosTest extends TestCase {
     private $profissional;
     private $paciente;
     private $senha;
@@ -49,26 +49,6 @@ class AgendamentosTest extends TestCase {
         $this->assertAuthenticated();
     }
 
-    private function loginPaciente() {
-        $resposta = $this->post(route('login'), [
-            'login' => $this->paciente->login,
-            'password' => $this->senha,
-            'remember' => "on",
-        ]);
-        $resposta->assertSessionHasNoErrors();
-        $resposta->assertRedirect(route('paciente.home'));
-        $resposta->assertLocation(route('paciente.home'));
-
-        Auth::login($this->paciente, true);
-        $this->assertTrue(Auth::check());
-        $this->assertAuthenticated();
-
-        $resposta = $this->get("/paciente/home");
-        $resposta->assertOk();
-
-        $resposta->assertSee("Informações Pessoais");
-    }
-
     private function logout() {
         $this->post(route('logout'));
         Auth::logout();
@@ -77,28 +57,11 @@ class AgendamentosTest extends TestCase {
     }
 
     /** @test **/
-    public function pacienteNaoPodeVerPaginaDeAgendamento() {
-        $this->loginPaciente();
-        $respota = $this->get(route("agendamento.criar"));
-        $respota->assertRedirect();
-        $respota->assertDontSee("Agendar para o paciente");
-    }
-
-    /** @test **/
     public function profissioanlPodeVerPaginaDeAgendamento() {
         $this->loginProfisssional();
         $respota = $this->get(route("agendamento.criar"));
         $respota->assertOk();
         $respota->assertSee("Agendar para o paciente");
-    }
-
-    /** @test **/
-    public function pacienteNaoPodeEnviarAgendamento() {
-        $this->loginPaciente();
-        $this->assertCount(0, Agendamentos::all());
-        $resposta = $this->post(route("agendamento.salvar"), $this->agendamento_completo);
-        $this->assertCount(0, Agendamentos::all());
-        $resposta->assertRedirect();
     }
 
     /** @test **/
@@ -488,7 +451,6 @@ class AgendamentosTest extends TestCase {
         $resposta = $this->post(route('agendamento.salvar'), $agendamento_incompleto);
         $resposta->assertSessionHasErrors('hora_saida');
         $this->assertCount(0, Agendamentos::all());
-
     }
 
 }
